Fix issue with tap file format
On Mac OS 11.3.1, running `mac_configure.sh` results in the following error:
```
Error: asana/pinned/python@2: Calling `sha256 "digest" => :tag` in a bottle block is disabled! Use `brew style --fix` on the formula to update the style or use `sha256 tag: "digest"` instead.
Please report this issue to the asana/pinned tap (not Homebrew/brew or Homebrew/core), or even better, submit a PR to fix it:
  /usr/local/Homebrew/Library/Taps/asana/homebrew-pinned/Formula/python@2.rb:11
```
This commit updates the format so that it will run successfully.